requirements.txt
0fastmcp>=3.0.2
httpx
pytz
markdown
reportlab
Pygments
openai